Rename v0_tests to v2_tests
