+++ tests/v2_tests/test_create_index_settings.py
