+++ tests/v2_tests/test_get_settings.py
