+++ tests/v2_tests/test_image_chunking.py
