+++ tests/v2_tests/test_index_init_logging.py
