+++ tests/v2_tests/test_logging.py
