+++ tests/v2_tests/test_model_auth.py
