+++ tests/v2_tests/test_telemetry.py
